Be more selective about what errors trigger missing dependency autoloads.
asdf-utils failed because of a missing dependency on ASDF, but it
wasn't just ASDF, but a particular version of ASDF. Quicklisp didn't
distinguish the error correctly and kept trying (and succeeding) to
load an ASDF that did not satisfy the dependency.

This is fixed in two ways.

First, handle the ASDF:MISSING-DEPENDENCY-OF-VERSION error, and
resignal it instead of trying to autoload the dependency it describes.

Second, keep track of which dependencies have been loaded already, and
signal an error if a loop appears.

Fixes https://github.com/quicklisp/quicklisp-projects/issues/352