Support installing non-primary systems not explicitly defined in .asd
When using ASDF's package-inferred-system, not all systems are
explicitly defined in the .asd file, and therefore do not get picked
up during the metadata generation. This makes it impossible to install
these subsystems directly (for example: "lil/pure/all").

This patch solves this by trying to find a system in a dist by its
primary name (the part before the first slash) if it is not found
verbatim. If the minimum ASDF version is upped in the future, I
recommend replacing much of this with `asdf:primary-system-name`.