Fix non-conforming code.
According to the LOAD-TIME-VALUE entry for CLHS
http://clhs.lisp.se/Body/s_ld_tim.htm, "It is guaranteed that the
evaluation of form will take place only once when the file is loaded,
but the order of evaluation with respect to the evaluation of top
level forms in the file is implementation-dependent." Therefore, doing
LOAD-TIME-VALUE of a function call where the function is defined in
the same file is not portable code.

This is important to fix because old versions of CMU CL do take
advantage of this fact, and future versions of SBCL may as well.