Handle symlinks to directories in LispWorks
If a symlink points to a directory, then LW will, if link transparency
is false (which it is) return its name as a filename from directory: as
a name whose file-namestring is not nil in other words.  Deal with this
in directory-entries by appending the name to the directory components.