Speed up constant database creation
There is no need to sync to disk every time a record is added, because
in a constant database records are only added at database creation time,
and the database is created in a temp file before being moved to the
final file. A sync is only needed once after all records have been
added. Further force-output is not a sync because it returns before
completion. finish-output is a sync (to the extent that Common Lisp
offers one).

With this patch, my systems.cdb gets created in 0.6 seconds instead of
44 seconds.